requirements.txt
redis
bcrypt
Flask-JWT-Extended
Flask
mysql-connector-python
python-dotenv
celery==5.5.3


debugpy
Flask-CORS
openmeteo-requests
requests-cache 
retry-requests 
numpy 
pandas

Flask-SocketIO
APScheduler
gevent
gevent-websocket
